requirements.txt
agentscope==1.0.18
fastapi==0.135.3
uvicorn==0.44.0
pydantic==2.12.5
gradio==6.13.0
sqlalchemy==2.0.49
python-jose[cryptography]>=3.5.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.13.1
httpx==0.28.1